Code Quality: Refactored ToolbarViewModel (#13659)
