+++ src/Files.App/Data/Items/PathBreadcrumbItem.cs
@@ -3,9 +3,9 @@
 
 namespace Files.App.Data.Items
 {
-	public class PathBoxItem
+	public class PathBreadcrumbItem
 	{
-		public string? Title { get; set; }
+		public string? Name { get; set; }
 
 		public string? Path { get; set; }
 	}
