+++ src/Files.App/Data/TemplateSelectors/PathBreadcrumbItemSelector.cs
@@ -15,21 +15,19 @@ namespace Files.App.Data.TemplateSelectors
 
 		public DataTemplate? CurrentItem { get; set; }
 
-		protected override DataTemplate SelectTemplateCore(object item, DependencyObject container)
+		protected override DataTemplate? SelectTemplateCore(object item, DependencyObject container)
 		{
 			var itemsControl = ItemsControl.ItemsControlFromItemContainer(container);
 
-			if (itemsControl.ItemsSource is ObservableCollection<PathBoxItem> items)
+			if (itemsControl.ItemsSource is ObservableCollection<PathBreadcrumbItem> items)
 			{
 				return
 					itemsControl.IndexFromContainer(container) == items.Count - 1
 						? CurrentItem!
 						: ParentItems!;
 			}
-			else
-			{
-				throw new ArgumentException($"Type of {nameof(itemsControl.ItemsSource)} doesn't match ObservableCollection<{nameof(PathBoxItem)}>");
-			}
+
+			return null;
 		}
 	}
 }
