+++ src/Files.App/UserControls/PathBreadcrumb.xaml.cs
@@ -7,50 +7,54 @@ using Microsoft.UI.Xaml.Input;
 
 namespace Files.App.UserControls
 {
-	[DependencyProperty<ToolbarViewModel>("ViewModel")]
+	/// <summary>
+	/// Represents control for breadcrumbs of item path.
+	/// </summary>
 	public sealed partial class PathBreadcrumb : UserControl
 	{
+		private PathBreadcrumbViewModel ViewModel { get; } = Ioc.Default.GetRequiredService<PathBreadcrumbViewModel>();
+
 		public PathBreadcrumb()
 		{
 			InitializeComponent();
 		}
 
-		private void PathItemSeparator_DataContextChanged(FrameworkElement sender, DataContextChangedEventArgs args)
+		private void PathBreadcrumbItemChevron_DataContextChanged(FrameworkElement sender, DataContextChangedEventArgs args)
 		{
 			ViewModel.PathItemSeparator_DataContextChanged(sender, args);
 		}
 
-		private void PathBoxItemFlyout_Opened(object sender, object e)
+		private void PathBreadcrumbItemFlyout_Opened(object sender, object e)
 		{
-			ViewModel.PathboxItemFlyout_Opened(sender, e);
+			ViewModel.PathBreadcrumbItemFlyout_Opened(sender, e);
 		}
 
-		private void PathBoxItem_DragLeave(object sender, DragEventArgs e)
+		private void PathBreadcrumbItem_DragLeave(object sender, DragEventArgs e)
 		{
-			ViewModel.PathBoxItem_DragLeave(sender, e);
+			ViewModel.PathBreadcrumbItem_DragLeave(sender, e);
 		}
 
-		private async void PathBoxItem_DragOver(object sender, DragEventArgs e)
+		private async void PathBreadcrumbItem_DragOver(object sender, DragEventArgs e)
 		{
-			await ViewModel.PathBoxItem_DragOver(sender, e);
+			await ViewModel.PathBreadcrumbItem_DragOver(sender, e);
 		}
 
-		private async void PathBoxItem_Drop(object sender, DragEventArgs e)
+		private async void PathBreadcrumbItem_Drop(object sender, DragEventArgs e)
 		{
-			await ViewModel.PathBoxItem_Drop(sender, e);
+			await ViewModel.PathBreadcrumbItem_Drop(sender, e);
 		}
 
-		private async void PathBoxItem_Tapped(object sender, TappedRoutedEventArgs e)
+		private async void PathBreadcrumbItem_Tapped(object sender, TappedRoutedEventArgs e)
 		{
-			await ViewModel.PathBoxItem_Tapped(sender, e);
+			await ViewModel.PathBreadcrumbItem_Tapped(sender, e);
 		}
 
-		private void PathBoxItem_PointerPressed(object sender, PointerRoutedEventArgs e)
+		private void PathBreadcrumbItem_PointerPressed(object sender, PointerRoutedEventArgs e)
 		{
-			ViewModel.PathBoxItem_PointerPressed(sender, e);
+			ViewModel.PathBreadcrumbItem_PointerPressed(sender, e);
 		}
 
-		private void PathItemSeparator_Loaded(object sender, RoutedEventArgs e)
+		private void PathBreadcrumbItemChevron_Loaded(object sender, RoutedEventArgs e)
 		{
 			if (sender is not FontIcon pathSeparatorIcon)
 				return;
