+++ src/Files.App/Views/HomePage.xaml.cs
@@ -229,16 +229,19 @@ namespace Files.App.Views
 			AppInstance.SlimContentPage?.DirectoryPropertiesViewModel.UpdateGitInfo(false, string.Empty, Array.Empty<BranchItem>());
 
 			// Clear the path UI and replace with Favorites
-			AppInstance.ToolbarViewModel.PathComponents.Clear();
-			string componentLabel = parameters.NavPathParam == "Home" ? "Home".GetLocalizedResource() : parameters.NavPathParam;
-			string tag = parameters.NavPathParam;
-			var item = new PathBoxItem()
+			AppInstance.PathBreadcrumbViewModel.PathBreadcrumbItems.Clear();
+
+			string componentLabel = parameters.NavPathParam == "Home" ? "Home".GetLocalizedResource() : parameters.NavPathParam ?? string.Empty;
+			string tag = parameters.NavPathParam ?? string.Empty;
+
+			var item = new PathBreadcrumbItem()
 			{
-				Title = componentLabel,
+				Name = componentLabel,
 				Path = tag,
 			};
 
-			AppInstance.ToolbarViewModel.PathComponents.Add(item);
+			AppInstance.PathBreadcrumbViewModel.PathBreadcrumbItems.Add(item);
+
 			base.OnNavigatedTo(eventArgs);
 		}
