+++ src/Files.App/Views/MainPage.xaml
@@ -179,7 +179,6 @@
 				HorizontalContentAlignment="Stretch"
 				x:Load="False"
 				Loaded="NavToolbar_Loaded"
-				OngoingTasksViewModel="{x:Bind OngoingTasksViewModel}"
 				ShowOngoingTasks="True"
 				ShowSearchBox="True"
 				ShowSettingsButton="{x:Bind WindowContext.IsCompactOverlay, Mode=OneWay, Converter={StaticResource BoolNegationConverter}}"
